Refactor: Creating user validations class
